Extract copyToClipboard function to utils.js and update usage across the codebase
