+++ addon/data-export.js
@@ -2,7 +2,8 @@
 import {sfConn, apiVersion} from "./inspector.js";
 import {getLinkTarget, nullToEmptyString, displayButton, PromptTemplate, Constants, UserInfoModel, createSpinForMethod} from "./utils.js";
 /* global initButton */
-import {Enumerable, DescribeInfo, copyToClipboard, initScrollTable, s} from "./data-load.js";
+import {copyToClipboard} from "./utils.js";
+import {Enumerable, DescribeInfo, initScrollTable, s} from "./data-load.js";
 import {PageHeader} from "./components/PageHeader.js";
 
 class QueryHistory {
