+++ addon/event-monitor.js
@@ -3,7 +3,7 @@ import {getLinkTarget, UserInfoModel, getBrowserType, createSpinForMethod} from 
 import {sfConn, apiVersion} from "./inspector.js";
 // Import the CometD library
 import {CometD} from "./lib/cometd/cometd.js";
-import {copyToClipboard} from "./data-load.js";
+import {copyToClipboard} from "./utils.js";
 import ConfirmModal from "./components/ConfirmModal.js";
 import {PageHeader} from "./components/PageHeader.js";
