+++ addon/popup.js
@@ -1,6 +1,6 @@
 /* global React ReactDOM */
 import {sfConn, apiVersion, sessionError} from "./inspector.js";
-import {getLinkTarget, displayButton, getLatestApiVersionFromOrg, setOrgInfo, getPKCEParameters, getBrowserType, getExtensionId, getClientId, getRedirectUri, Constants} from "./utils.js";
+import {getLinkTarget, displayButton, getLatestApiVersionFromOrg, setOrgInfo, getPKCEParameters, getBrowserType, getExtensionId, getClientId, getRedirectUri, Constants, copyToClipboard} from "./utils.js";
 import {setupLinks} from "./links.js";
 import AlertBanner from "./components/AlertBanner.js";
 
@@ -3312,11 +3312,22 @@ class UserDetails extends React.PureComponent {
                     title: "Show all data",
                   },
                   user.Id,
-                  h("span", {title: "Copy to clipboard", onClick: (e) => handleUserIdCopy(e, user.Id)},
-                    h("svg", {className: "slds-button__icon slds-m-left_xx-small", style: "vertical-align: sub;"},
+                  displayButton("copy-userId", hideButtonsOption)
+                    ? h("span", {
+                      className: "sfir-copy-userid-icon",
+                      title: "Copy to clipboard",
+                      onClick: (e) => {
+                        e.stopPropagation();
+                        handleUserIdCopy(e, user.Id);
+                      },
+                      onMouseEnter: (e) => e.stopPropagation(),
+                      onMouseLeave: (e) => e.stopPropagation()
+                    },
+                    h("svg", {className: "slds-button__icon slds-m-left_xx-small sfir-vertical-align_sub"},
                       h("use", {xlinkHref: "symbols.svg#copy"})
                     )
-                  ),
+                    )
+                    : null,
                 )
               )
             ),
@@ -4755,27 +4766,5 @@ function handleLightningLinkClick(e) {
 
 function handleUserIdCopy(e, userId) {
   e.preventDefault();
-
-  // Create a temporary input element to trigger the oncopy event
-  const tempInput = document.createElement("input");
-  tempInput.value = "temp";
-
-  // Set up the copy event to override clipboard data
-  tempInput.addEventListener("copy", (event) => {
-    event.clipboardData.setData("text/plain", userId);
-    event.preventDefault();
-  });
-
-  document.body.appendChild(tempInput);
-
-  try {
-    tempInput.select();
-    // Simulate a copy programatically
-    const success = document.execCommand("copy");
-    if (!success) {
-      alert("Copy failed");
-    }
-  } finally {
-    document.body.removeChild(tempInput);
-  }
+  copyToClipboard(userId);
 }
